Create Main_Menu
This implementation starts with the option  and when S1 is pressed nothing happenes. When instead S2 is pressed it jumps to Read and if pressed again to Log. When we press S2 continuously it just cycles between Log and Read. If press S1 in wither of these two options it would just do a reti.